refactor: 테스트를 실제 패키지 구조와 동일하도록 수정
