--- src/test/java/boardSetting/PieceTest/PawnTest.java
@@ -1,4 +1,0 @@
-package boardSetting.PieceTest;
-
-public class PawnTest {
-}
